requirements.txt
streamlit
langchain
flake8
gitpython
sqlalchemy
python-dotenv
pydantic
riva-api
nemo_toolkit[all] #all parts
transformers
requests
